requirements.txt
streamlit
edge-tts
pydub
SpeechRecognition
deep-translator
httpx
Pillow